requirements.txt
-i https://pypi.org/simple
blinker==1.9.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
flask==3.1.2; python_version >= '3.9'
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
pymysql==1.1.0
flask-mysqldb==1.0.1
python-dotenv==1.0.0
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
